feat(v8): throw proper DOMException for WebIDL errors
Add throwDOMException() and throwWebIDLError() to conversions.zig that
properly create DOMException instances using Reflect.construct instead
of generic Error objects.

Update all error handling in interface.zig to use throwWebIDLError()
which checks if the error name is a known DOMException name and throws
the appropriate exception type.

This fixes WPT tests that use assert_throws_dom() which check:
- e instanceof DOMException === true
- e.name === 'InvalidStateError' (or other DOMException names)

The InvalidStateError that was appearing as a generic Error will now
be properly thrown as a DOMException with the correct name property.