fix(webidl): cache NamedNodeMap in Element.get_attributes()
Per WebIDL semantics, the attributes getter must return the same object
on subsequent accesses so that own properties added to it persist.

Previously, get_attributes() created a new NamedNodeMap on every call,
which meant:
  element.attributes.foo = 'bar';
  element.attributes.foo; // undefined (different object!)

Now the NamedNodeMap is cached in Element's InternalState and returned
on subsequent accesses. Proper cleanup is handled in deinit.

Fixes: whatwg-bwqn1